Fått fjernet klassen TestContext
