+++ app/src/test/kotlin/no/nav/k9punsj/rest/eksternt/punsjbollen/TestPunsjbolleService.kt
@@ -1,0 +1,39 @@
+package no.nav.k9punsj.rest.eksternt.punsjbollen
+
+import no.nav.k9.søknad.Søknad
+import no.nav.k9punsj.CorrelationId
+import no.nav.k9punsj.TestProfil
+import no.nav.k9punsj.rest.web.dto.JournalpostIdDto
+import no.nav.k9punsj.rest.web.dto.NorskIdentDto
+import no.nav.k9punsj.rest.web.dto.PeriodeDto
+import no.nav.k9punsj.rest.web.dto.SaksnummerDto
+import org.springframework.stereotype.Component
+
+@Component
+@TestProfil
+internal class TestPunsjbolleService : PunsjbolleService {
+    override suspend fun opprettEllerHentFagsaksnummer(
+        søker: NorskIdentDto,
+        barn: NorskIdentDto,
+        journalpostId: JournalpostIdDto?,
+        periode: PeriodeDto?,
+        correlationId: CorrelationId
+    ) = require(journalpostId != null || periode != null) {
+        "Må sette minst en av journalpostId og periode"
+    }.let { SaksnummerDto("133742069666") }
+
+    override suspend fun opprettEllerHentFagsaksnummer(
+        søker: NorskIdentDto,
+        barn: NorskIdentDto,
+        søknad: Søknad,
+        correlationId: CorrelationId
+    ) = SaksnummerDto("133742069666")
+
+    override suspend fun ruting(
+        søker: NorskIdentDto,
+        barn: NorskIdentDto,
+        journalpostId: JournalpostIdDto?,
+        periode: PeriodeDto?,
+        correlationId: CorrelationId
+    ) = PunsjbolleRuting.K9Sak
+}
